backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
anthropic==0.7.0
pypdf==3.17.1
python-docx==1.1.0
python-dotenv==1.0.0
